fix: widen the TTY to avoid truncating commands
The default size of the pexpect TTY is 24 high x 80 wide. If a command
gets close to 80 chars, it gets truncated (I think by bash's line
editing prompt), which causes the .expect() calls matching the command
to hang and timeout.

To fix this, the TTY we execute commands in is now 1000 chars wide. So
commands can be close to 1000 chars long before they get truncated,
which should be much more than any reasonable documentation example.

If longer commands are required, we can work around by expect()ing just
a fixed-length tail of the command rather than the whole thing. The
actual output of the program does not get truncated, regardless of TTY
width.